More clean up - Move ChannelReader conversion logic to helper method. - Remove extra left over observable methods and classes
